MAR-179 Complete user profile, settings, and account deletion flow (Apple Guideline 5.1.1 compliance)
