+++ Marko/Features/Account/ProfileVM.swift
@@ -21,9 +21,6 @@ class ProfileVM {
     
     var didFinishAction: (() -> ())?
     
-    var didProfileTapped: (() -> ())?
-
-    
     // MARK: - Methods
     
     func signOut() {
