+++ Marko/Features/TeacherDetail/TeacherDetailVC.swift
@@ -13,9 +13,6 @@ import PassKit
 class TeacherDetailVC: UIViewController {
     
     private let vm: TeacherDetailVM
-    
-    var profileVM: ProfileVM?
-    
 
     private var selectedSlotToBook: TimeSlot?
     
@@ -84,22 +81,6 @@ class TeacherDetailVC: UIViewController {
         stack.spacing = 10
         return stack
     }()
-    
-    lazy private var profileButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.setTitle("Profile", for: .normal)
-        button.backgroundColor = .systemGray3
-        button.setTitleColor(.white, for: .normal)
-        button.titleLabel?.font = UIFont.systemFont(ofSize: 14, weight: .bold)
-        button.layer.cornerRadius = 15
-        button.addAction(UIAction { _ in
-            
-            self.profileVM?.didProfileTapped?()
-            
-        }, for: .touchUpInside)
-        
-        return button
-    }()
 
     init(vm: TeacherDetailVM) {
         self.vm = vm
@@ -120,7 +101,6 @@ class TeacherDetailVC: UIViewController {
         setupViews()
         setupLayout()
         bindTeacherData()
-        setupNavBar()
         
         calendarView.delegate = self
         
@@ -189,13 +169,6 @@ class TeacherDetailVC: UIViewController {
         calendarView.snp.makeConstraints { make in
             make.height.equalTo(300)
         }
-    }
-    
-    private func setupNavBar() {
-        
-        let button = UIBarButtonItem(customView: profileButton)
-        navigationItem.rightBarButtonItem = button
-        
     }
     
     // MARK: - Booking UI
